feat: unify log viewer components with enhanced functionality
- Create UnifiedLogViewer component combining real-time and historical log viewing
- Support both 'realtime' mode for current translation and 'historical' mode for past sessions
- Enhanced historical log viewer with better formatting, filtering, and larger width
- Maintain all advanced features from real-time viewer (auto-scroll, filtering, styling)
- Replace simple text-based historical log viewer with full-featured component
- Backward compatibility maintained through wrapper components
