+++ src/components/ui/log-dialog.tsx
@@ -1,26 +1,5 @@
-import React, { useState, useEffect, useRef, useCallback } from 'react';
-import { Dialog, DialogContent, DialogHeader, DialogTitle, DialogFooter } from './dialog';
-import { Button } from './button';
-import { Card } from './card';
-import { ScrollArea } from './scroll-area';
-import { useAppTranslation } from '@/lib/i18n';
-import { useAppStore } from '@/lib/store';
-import { FileService } from '@/lib/services/file-service';
-import { listen } from '@tauri-apps/api/event';
-import { UI_DEFAULTS } from '@/lib/constants/defaults';
-
-// Log entry type
-interface LogEntry {
-  timestamp: string;
-  level: {
-    Debug: null;
-    Info: null;
-    Warning: null;
-    Error: null;
-  } | string;
-  message: string;
-  process_type?: string;
-}
+import React from 'react';
+import { UnifiedLogViewer } from './unified-log-viewer';
 
 interface LogDialogProps {
   open: boolean;
@@ -28,673 +7,14 @@ interface LogDialogProps {
 }
 
 export function LogDialog({ open, onOpenChange }: LogDialogProps) {
-  const { t } = useAppTranslation();
-  const isTranslating = useAppStore((state) => state.isTranslating);
-  const [autoScroll, setAutoScroll] = useState(true);
-  const [logs, setLogs] = useState<LogEntry[]>([]);
-  const [userInteracting, setUserInteracting] = useState(false);
-  const scrollViewportRef = useRef<HTMLDivElement>(null);
-  const interactionTimeoutRef = useRef<NodeJS.Timeout | null>(null);
-  
-  // Callback ref to get the viewport element from ScrollArea
-  const scrollAreaCallbackRef = useCallback((element: HTMLDivElement | null) => {
-    if (element) {
-      // Find the viewport element within the ScrollArea
-      const viewport = element.querySelector('[data-slot="scroll-area-viewport"]') as HTMLDivElement;
-      if (viewport) {
-        scrollViewportRef.current = viewport;
-      }
-    }
-  }, []);
-  
-  // Function to get log level string
-  const getLogLevelString = (level: LogEntry['level']): string => {
-    if (typeof level === 'string') {
-      return level;
-    }
-    
-    if ('Error' in level) return 'ERROR';
-    if ('Warning' in level) return 'WARNING';
-    if ('Info' in level) return 'INFO';
-    if ('Debug' in level) return 'DEBUG';
-    
-    return 'UNKNOWN';
-  };
-  
-  // Function to get log level color
-  const getLogLevelColor = (level: LogEntry['level']) => {
-    const levelStr = getLogLevelString(level).toLowerCase();
-    
-    switch (levelStr) {
-      case 'error':
-        return 'text-red-500 dark:text-red-400';
-      case 'warning':
-      case 'warn':
-        return 'text-yellow-500 dark:text-yellow-400';
-      case 'info':
-        return 'text-blue-500 dark:text-blue-400';
-      case 'debug':
-        return 'text-gray-500 dark:text-gray-400';
-      default:
-        return 'text-gray-700 dark:text-gray-300';
-    }
-  };
-  
-  // Function to format log message
-  const formatLogMessage = (log: LogEntry) => {
-    let message = '';
-    
-    // Always show timestamp and level in the dialog
-    message += `[${log.timestamp}] `;
-    message += `[${getLogLevelString(log.level)}] `;
-    
-    // Process type is optional
-    if (log.process_type) {
-      message += `[${log.process_type}] `;
-    }
-    
-    message += log.message;
-    
-    return message;
-  };
-  
-  // Function to filter logs - only show important logs in the dialog
-  const filterLogs = (logs: LogEntry[]) => {
-    return logs.filter(log => {
-      const levelStr = getLogLevelString(log.level).toLowerCase();
-      
-      // Only show logs that are important for the user to see
-      // 1. All error logs
-      if (levelStr === 'error') {
-        return true;
-      }
-      
-      // 1.5. Show ErrorLogger messages (they contain important context)
-      if (log.message.includes('[ErrorLogger]') || log.message.includes('TranslationService.logError')) {
-        return true;
-      }
-      
-      // During active translation, show more logs
-      if (isTranslating) {
-        // Show all translation-related logs during translation
-        if (log.process_type === 'TRANSLATION' || 
-            log.process_type === 'TRANSLATION_START' ||
-            log.process_type === 'TRANSLATION_STATS' ||
-            log.process_type === 'TRANSLATION_PROGRESS' ||
-            log.process_type === 'TRANSLATION_COMPLETE') {
-          return true;
-        }
-        
-        // Show all API request logs during translation
-        if (log.process_type === 'API_REQUEST') {
-          return true;
-        }
-        
-        // Show file operations during translation
-        if (log.process_type === 'FILE_OPERATION') {
-          return true;
-        }
-        
-        // Show warnings during translation
-        if (levelStr === 'warning' || levelStr === 'warn') {
-          return true;
-        }
-        
-        // Show info logs during translation
-        if (levelStr === 'info') {
-          return true;
-        }
-      }
-      
-      // When not translating, only show critical logs
-      // 2. Enhanced translation process logs
-      if (log.process_type === 'TRANSLATION') {
-        // Filter out verbose translation logs that aren't useful to users
-        const message = log.message.toLowerCase();
-        // Skip detailed chunk processing logs unless they're errors
-        if (message.includes('chunk') && !message.includes('error') && !message.includes('failed')) {
-          return false;
-        }
-        return true;
-      }
-      
-      // 3. New enhanced translation logging categories
-      if (log.process_type === 'TRANSLATION_START') {
-        return true; // Always show translation start logs
-      }
-      
-      if (log.process_type === 'TRANSLATION_STATS') {
-        return true; // Show pre-translation statistics
-      }
-      
-      if (log.process_type === 'TRANSLATION_PROGRESS') {
-        // Show file progress but limit frequency to avoid spam
-        const message = log.message.toLowerCase();
-        // Only show progress at certain milestones or completion
-        if (message.includes('100%') || message.includes('completed') || 
-            message.includes('50%') || message.includes('75%')) {
-          return true;
-        }
-        return false;
-      }
-      
-      if (log.process_type === 'TRANSLATION_COMPLETE') {
-        return true; // Always show completion summaries
-      }
-      
-      // 4. Performance logs - only show for errors or important milestones
-      if (log.process_type === 'PERFORMANCE') {
-        // Only show performance logs for debug purposes (can be filtered out in production)
-        return levelStr === 'error' || levelStr === 'warning';
-      }
-      
-      // 5. API request logs
-      if (log.process_type === 'API_REQUEST') {
-        return true;
-      }
-      
-      // 6. File operation logs
-      if (log.process_type === 'FILE_OPERATION') {
-        return true;
-      }
-      
-      // 7. System logs
-      if (log.process_type === 'SYSTEM') {
-        return true;
-      }
-      
-      // 8. Warnings that might be important
-      if (levelStr === 'warning' || levelStr === 'warn') {
-        return true;
-      }
-      
-      // Filter out debug and info logs that aren't important for users
-      return false;
-    });
-  };
-  
-  // Note: We don't reset logs when translation starts anymore
-  // This allows users to see logs from previous translations
-  
-  // Effect to listen for log events from Tauri
-  useEffect(() => {
-    // Skip in SSR
-    if (typeof window === 'undefined') return;
-    
-    // Function to load initial logs
-    const loadInitialLogs = async () => {
-      try {
-        // Check if we're in a Tauri environment
-        if (typeof window !== 'undefined' && typeof (window as unknown as Record<string, unknown>).__TAURI_INTERNALS__ !== 'undefined') {
-          // Use FileService to invoke the get_logs command
-          const initialLogs = await FileService.invoke<LogEntry[]>('get_logs');
-          console.log('[LogDialog] Initial logs loaded:', initialLogs);
-          setLogs(initialLogs || []);
-        }
-      } catch (error) {
-        console.error('Failed to load initial logs:', error);
-      }
-    };
-    
-    // Function to listen for log events
-    const listenForLogs = async () => {
-      try {
-        // Check if we're in a Tauri environment
-        if (typeof window !== 'undefined' && typeof (window as unknown as Record<string, unknown>).__TAURI_INTERNALS__ !== 'undefined') {
-          // Listen for log events using Tauri v2 API
-          const unlistenFn = await listen<LogEntry>('log', (event) => {
-            console.log('[LogDialog] Received log event:', event.payload);
-            setLogs(prevLogs => [...prevLogs, event.payload]);
-          });
-          
-          // Return cleanup function
-          return unlistenFn;
-        }
-      } catch (error) {
-        console.error('Failed to listen for log events:', error);
-      }
-      
-      // Return no-op cleanup function
-      return () => {};
-    };
-    
-    // Load initial logs
-    loadInitialLogs();
-    
-    // Listen for log events
-    const unlistenPromise = listenForLogs();
-    
-    // Cleanup
-    return () => {
-      unlistenPromise.then(unlisten => unlisten && unlisten());
-    };
-  }, []);
-  
-  // Handle user interaction detection
-  const handleUserScroll = () => {
-    setUserInteracting(true);
-    
-    // Clear existing timeout
-    if (interactionTimeoutRef.current) {
-      clearTimeout(interactionTimeoutRef.current);
-    }
-    
-    // Set a timeout to mark interaction as finished after 2 seconds
-    interactionTimeoutRef.current = setTimeout(() => {
-      setUserInteracting(false);
-    }, UI_DEFAULTS.autoScroll.interactionDelay);
-  };
-
-  // Effect to auto-scroll to bottom (only when not actively interacting)
-  useEffect(() => {
-    if (autoScroll && !userInteracting && scrollViewportRef.current) {
-      const viewport = scrollViewportRef.current;
-      viewport.scrollTop = viewport.scrollHeight;
-    }
-  }, [logs, autoScroll, userInteracting]);
-
-  // Cleanup timeout on unmount
-  useEffect(() => {
-    return () => {
-      if (interactionTimeoutRef.current) {
-        clearTimeout(interactionTimeoutRef.current);
-      }
-    };
-  }, []);
-  
-  // Keep dialog open for a few seconds after translation completes to show final logs
-  useEffect(() => {
-    if (!isTranslating && open) {
-      // Keep the dialog open for a few seconds after translation completes
-      const timer = setTimeout(() => {
-        // Don't auto-close if there was an error
-        if (!useAppStore.getState().error) {
-          onOpenChange(false);
-        }
-      }, UI_DEFAULTS.dialog.autoCloseDelay);
-      
-      return () => clearTimeout(timer);
-    }
-  }, [isTranslating, open, onOpenChange]);
-  
-  // Filter logs
-  const filteredLogs = filterLogs(logs);
-  
   return (
-    <Dialog open={open} onOpenChange={onOpenChange}>
-      <DialogContent className="sm:max-w-[900px] 2xl:max-w-[1400px] max-h-[80vh] 2xl:max-h-[85vh]">
-        <DialogHeader>
-          <DialogTitle>{t('logs.translationLogs')}</DialogTitle>
-        </DialogHeader>
-        
-        <div className="py-4">
-          <ScrollArea 
-            ref={scrollAreaCallbackRef}
-            className="border rounded-md"
-            style={{ height: UI_DEFAULTS.scrollArea.defaultHeight }}
-            onScroll={handleUserScroll}
-            onWheel={handleUserScroll}
-            onMouseDown={handleUserScroll}
-            onTouchStart={handleUserScroll}
-          >
-            <div className="p-4 2xl:p-6 font-mono text-sm 2xl:text-base whitespace-pre-wrap">
-              {filteredLogs.length === 0 ? (
-                <div className="text-gray-500 dark:text-gray-400">
-                  {t('logs.noLogs')}
-                </div>
-              ) : (
-                filteredLogs.map((log, index) => (
-                  <div 
-                    key={`${log.timestamp}-${index}`}
-                    className={`${getLogLevelColor(log.level)} mb-1`}
-                  >
-                    {formatLogMessage(log)}
-                  </div>
-                ))
-              )}
-            </div>
-          </ScrollArea>
-        </div>
-        
-        <DialogFooter className="flex justify-between items-center">
-          <div className="flex items-center space-x-2">
-            <input
-              type="checkbox"
-              id="auto-scroll"
-              checked={autoScroll}
-              onChange={(e) => setAutoScroll(e.target.checked)}
-              className="rounded border-gray-300 text-primary-600 focus:ring-primary-500"
-            />
-            <label htmlFor="auto-scroll" className="text-sm">
-              {t('logs.autoScroll')}
-            </label>
-          </div>
-          
-          <div className="flex gap-2">
-            <Button 
-              variant="outline"
-              onClick={async () => {
-                try {
-                  await FileService.invoke('clear_logs');
-                  setLogs([]);
-                  console.log('[LogDialog] Logs cleared');
-                } catch (error) {
-                  console.error('[LogDialog] Failed to clear logs:', error);
-                }
-              }}
-            >
-              Clear Logs
-            </Button>
-            <Button onClick={() => onOpenChange(false)}>
-              {t('common.close')}
-            </Button>
-          </div>
-        </DialogFooter>
-      </DialogContent>
-    </Dialog>
+    <UnifiedLogViewer
+      open={open}
+      onOpenChange={onOpenChange}
+      mode="realtime"
+    />
   );
 }
 
-/**
- * Standalone log viewer component for use in other parts of the application
- * This is a wrapper around the log functionality to be used outside of the dialog
- */
-interface LogViewerProps {
-  height?: string;
-  autoScroll?: boolean;
-  showTimestamp?: boolean;
-  showLevel?: boolean;
-  showSource?: boolean;
-  filter?: string;
-}
-
-export function LogViewer({
-  height = UI_DEFAULTS.scrollArea.logViewerHeight,
-  autoScroll = true,
-  showTimestamp = true,
-  showLevel = true,
-  showSource = false,
-  filter
-}: LogViewerProps) {
-  const { t } = useAppTranslation();
-  const isTranslating = useAppStore((state) => state.isTranslating);
-  const [logs, setLogs] = useState<LogEntry[]>([]);
-  const [userInteracting, setUserInteracting] = useState(false);
-  const scrollViewportRef2 = useRef<HTMLDivElement>(null);
-  const interactionTimeoutRef = useRef<NodeJS.Timeout | null>(null);
-  
-  // Callback ref to get the viewport element from ScrollArea
-  const scrollAreaCallbackRef2 = useCallback((element: HTMLDivElement | null) => {
-    if (element) {
-      // Find the viewport element within the ScrollArea
-      const viewport = element.querySelector('[data-slot="scroll-area-viewport"]') as HTMLDivElement;
-      if (viewport) {
-        scrollViewportRef2.current = viewport;
-      }
-    }
-  }, []);
-  
-  // Function to get log level string
-  const getLogLevelString = (level: LogEntry['level']): string => {
-    if (typeof level === 'string') {
-      return level;
-    }
-    
-    if ('Error' in level) return 'ERROR';
-    if ('Warning' in level) return 'WARNING';
-    if ('Info' in level) return 'INFO';
-    if ('Debug' in level) return 'DEBUG';
-    
-    return 'UNKNOWN';
-  };
-  
-  // Function to get log level color
-  const getLogLevelColor = (level: LogEntry['level']) => {
-    const levelStr = getLogLevelString(level).toLowerCase();
-    
-    switch (levelStr) {
-      case 'error':
-        return 'text-red-500 dark:text-red-400';
-      case 'warning':
-      case 'warn':
-        return 'text-yellow-500 dark:text-yellow-400';
-      case 'info':
-        return 'text-blue-500 dark:text-blue-400';
-      case 'debug':
-        return 'text-gray-500 dark:text-gray-400';
-      default:
-        return 'text-gray-700 dark:text-gray-300';
-    }
-  };
-  
-  // Function to format log message
-  const formatLogMessage = (log: LogEntry) => {
-    let message = '';
-    
-    if (showTimestamp) {
-      message += `[${log.timestamp}] `;
-    }
-    
-    if (showLevel) {
-      message += `[${getLogLevelString(log.level)}] `;
-    }
-    
-    if (showSource && log.process_type) {
-      message += `[${log.process_type}] `;
-    }
-    
-    message += log.message;
-    
-    return message;
-  };
-  
-  // Function to filter logs - only show important logs in the viewer
-  const filterLogs = (logs: LogEntry[]) => {
-    // First apply the custom filter if provided
-    let filteredByCustom = logs;
-    if (filter) {
-      filteredByCustom = logs.filter(log => {
-        const message = formatLogMessage(log).toLowerCase();
-        return message.includes(filter.toLowerCase());
-      });
-    }
-    
-    // Then apply the importance filter
-    return filteredByCustom.filter(log => {
-      const levelStr = getLogLevelString(log.level).toLowerCase();
-      
-      // Only show logs that are important for the user to see
-      // 1. All error logs
-      if (levelStr === 'error') {
-        return true;
-      }
-      
-      // 2. Enhanced translation process logs
-      if (log.process_type === 'TRANSLATION') {
-        // Filter out verbose translation logs that aren't useful to users
-        const message = log.message.toLowerCase();
-        // Skip detailed chunk processing logs unless they're errors
-        if (message.includes('chunk') && !message.includes('error') && !message.includes('failed')) {
-          return false;
-        }
-        return true;
-      }
-      
-      // 3. New enhanced translation logging categories
-      if (log.process_type === 'TRANSLATION_START') {
-        return true; // Always show translation start logs
-      }
-      
-      if (log.process_type === 'TRANSLATION_STATS') {
-        return true; // Show pre-translation statistics
-      }
-      
-      if (log.process_type === 'TRANSLATION_PROGRESS') {
-        // Show file progress but limit frequency to avoid spam
-        const message = log.message.toLowerCase();
-        // Only show progress at certain milestones or completion
-        if (message.includes('100%') || message.includes('completed') || 
-            message.includes('50%') || message.includes('75%')) {
-          return true;
-        }
-        return false;
-      }
-      
-      if (log.process_type === 'TRANSLATION_COMPLETE') {
-        return true; // Always show completion summaries
-      }
-      
-      // 4. Performance logs - only show for errors or important milestones
-      if (log.process_type === 'PERFORMANCE') {
-        // Only show performance logs for debug purposes (can be filtered out in production)
-        return levelStr === 'error' || levelStr === 'warning';
-      }
-      
-      // 5. API request logs
-      if (log.process_type === 'API_REQUEST') {
-        return true;
-      }
-      
-      // 6. File operation logs
-      if (log.process_type === 'FILE_OPERATION') {
-        return true;
-      }
-      
-      // 7. System logs
-      if (log.process_type === 'SYSTEM') {
-        return true;
-      }
-      
-      // 8. Warnings that might be important
-      if (levelStr === 'warning' || levelStr === 'warn') {
-        return true;
-      }
-      
-      // Filter out debug and info logs that aren't important for users
-      return false;
-    });
-  };
-  
-  // Effect to listen for log events from Tauri
-  useEffect(() => {
-    // Skip in SSR
-    if (typeof window === 'undefined') return;
-    
-    // Function to load initial logs
-    const loadInitialLogs = async () => {
-      try {
-        // Check if we're in a Tauri environment
-        if (typeof window !== 'undefined' && typeof (window as unknown as Record<string, unknown>).__TAURI_INTERNALS__ !== 'undefined') {
-          // Use FileService to invoke the get_logs command
-          const initialLogs = await FileService.invoke<LogEntry[]>('get_logs');
-          console.log('[LogDialog] Initial logs loaded:', initialLogs);
-          setLogs(initialLogs || []);
-        }
-      } catch (error) {
-        console.error('Failed to load initial logs:', error);
-      }
-    };
-    
-    // Function to listen for log events
-    const listenForLogs = async () => {
-      try {
-        // Check if we're in a Tauri environment
-        if (typeof window !== 'undefined' && typeof (window as unknown as Record<string, unknown>).__TAURI_INTERNALS__ !== 'undefined') {
-          // Listen for log events using Tauri v2 API
-          const unlistenFn = await listen<LogEntry>('log', (event) => {
-            console.log('[LogDialog] Received log event:', event.payload);
-            setLogs(prevLogs => [...prevLogs, event.payload]);
-          });
-          
-          // Return cleanup function
-          return unlistenFn;
-        }
-      } catch (error) {
-        console.error('Failed to listen for log events:', error);
-      }
-      
-      // Return no-op cleanup function
-      return () => {};
-    };
-    
-    // Load initial logs
-    loadInitialLogs();
-    
-    // Listen for log events
-    const unlistenPromise = listenForLogs();
-    
-    // Cleanup
-    return () => {
-      unlistenPromise.then(unlisten => unlisten && unlisten());
-    };
-  }, []);
-  
-  // Handle user interaction detection for LogViewer
-  const handleUserScrollViewer = () => {
-    setUserInteracting(true);
-    
-    // Clear existing timeout
-    if (interactionTimeoutRef.current) {
-      clearTimeout(interactionTimeoutRef.current);
-    }
-    
-    // Set a timeout to mark interaction as finished after 2 seconds
-    interactionTimeoutRef.current = setTimeout(() => {
-      setUserInteracting(false);
-    }, UI_DEFAULTS.autoScroll.interactionDelay);
-  };
-
-  // Effect to auto-scroll to bottom (only when not actively interacting)
-  useEffect(() => {
-    if (autoScroll && !userInteracting && scrollViewportRef2.current) {
-      const viewport = scrollViewportRef2.current;
-      viewport.scrollTop = viewport.scrollHeight;
-    }
-  }, [logs, autoScroll, userInteracting]);
-
-  // Cleanup timeout on unmount for LogViewer
-  useEffect(() => {
-    return () => {
-      if (interactionTimeoutRef.current) {
-        clearTimeout(interactionTimeoutRef.current);
-      }
-    };
-  }, []);
-  
-  // Filter logs
-  const filteredLogs = filterLogs(logs);
-  
-  return (
-    <Card className="w-full">
-      <div className="p-4">
-        <h3 className="text-lg font-medium">{t('logs.title')}</h3>
-        <ScrollArea 
-          ref={scrollAreaCallbackRef2}
-          className="mt-2 border rounded-md"
-          style={{ height }}
-          onScroll={handleUserScrollViewer}
-          onWheel={handleUserScrollViewer}
-          onMouseDown={handleUserScrollViewer}
-          onTouchStart={handleUserScrollViewer}
-        >
-          <div className="p-4 font-mono text-sm whitespace-pre-wrap">
-            {filteredLogs.length === 0 ? (
-              <div className="text-gray-500 dark:text-gray-400">
-                {t('logs.noLogs')}
-              </div>
-            ) : (
-              filteredLogs.map((log, index) => (
-                <div 
-                  key={`${log.timestamp}-${index}`}
-                  className={`${getLogLevelColor(log.level)} mb-1`}
-                >
-                  {formatLogMessage(log)}
-                </div>
-              ))
-            )}
-          </div>
-        </ScrollArea>
-      </div>
-    </Card>
-  );
-}
+// Re-export the UnifiedLogViewer for backward compatibility
+export { UnifiedLogViewer } from './unified-log-viewer';
